requirements.txt
faker==40.11.1
pandas==2.2.2
pyarrow==15.0.2
google-cloud-storage==2.16.0
gcsfs==2024.3.1
pytest==8.1.1